+++ client/src/app/(withauth)/activity/page.tsx
@@ -6,15 +6,13 @@ import Show from "@/components/show-component"
 import { endPoints } from "@/constants/api-endpoints"
 import HTTPMethods from "@/constants/http-methods"
 import useFetch from "@/hooks/use-fetch"
-import { BellIcon, DashboardIcon } from "@radix-ui/react-icons"
+import { BellIcon } from "@radix-ui/react-icons"
 import moment from "moment"
-import { useRouter } from "next/navigation"
 import { Fragment, useCallback } from "react"
 import { Container } from "react-bootstrap"
 
 export default function Page() {
-  const router = useRouter()
-  const activities = useFetch("get-activities", endPoints.getAllActivities, HTTPMethods.GET, {}, true)
+  const activities = useFetch("get-activities", endPoints.getAllActivities, HTTPMethods.GET)
 
   const displayActivities = useCallback(() => {
     const activitiesToDisplay = activities?.data?.activities?.slice(0, 6).map((activity: any) => {