+++ client/src/app/(withauth)/products/nftstudio/page.tsx
@@ -13,18 +13,17 @@ import moment from "moment"
 import HTTPMethods from "@/constants/http-methods"
 import useFetch from "@/hooks/use-fetch"
 import { PlusCircledIcon } from "@radix-ui/react-icons"
-import ProductCard from "@/components/productcard-component"
-import { ProductCardInterface } from "@/types/Types"
+import ProductCard, { ProductCardInterface } from "@/components/productcard-component"
 import Hero from "@/components/hero-component"
 import { uiConstants } from "@/constants/global-constants"
 
 export default function Page() {
   const nftContractAddress = useFetch("secret-config", endPoints.nftstudioGetContractAddress, HTTPMethods.GET)
+  const products = useFetch("get-products", `${endPoints.getProductConfig}?searchQuery=nftstudio`, HTTPMethods.GET)
   const web3Provider = new Web3(endPoints.nftstudioSignTransactionGateway)
   const [{ userState, globalSearchString }] = useContext(GlobalContext)
   const [nftList, setNFTList] = useState([])
   const [isLoading, setLoading] = useState(false)
-  const products = useFetch("get-products", `${endPoints.getProductConfig}?searchQuery=nftstudio`, HTTPMethods.GET)
   const selectedProduct = products?.data?.find((product: any) => product.productName === "nftstudio")
 
   useEffect(() => {