+++ client/src/components/show-component.tsx
@@ -1,6 +1,10 @@
 "use client"
-import { ShowProps } from "@/types/Types"
-import { Fragment } from "react"
+import { Fragment, ReactNode } from "react"
+
+export interface ShowProps {
+  when: boolean
+  children: ReactNode
+}
 
 export default function Show({ when, children }: ShowProps) {
   return when === undefined || !when ? <Fragment></Fragment> : <Fragment>{children}</Fragment>